+++ src/main/java/teamport/aether/mixin/armor/player/phoenix/PlayerMixinFireImmunity.java
@@ -14,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import teamport.aether.api.ContainerHelper;
-import teamport.aether.api.ParticalHelper;
+import teamport.aether.particle.ParticalHelper;
 import teamport.aether.items.AetherArmorMaterial;
 import teamport.aether.mixin.accessors.EntityAccessor;
