+++ src/main/java/teamport/aether/mixin/armor/wolf/MobWolfMixinFireImmunity.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import teamport.aether.api.ParticalHelper;
+import teamport.aether.particle.ParticalHelper;
 import teamport.aether.items.AetherArmorMaterial;
 
 @Mixin(value = MobWolf.class, remap = false)
